Book_Inventary
Simple project where the system maintains the inventory of books in a bookshop. If a customer purchases a book, the book's count will decrease; if a book is added, the same is updated.You can add Book  and Book ID and make the search based on book ID or make the search using just one parameter giving multiple results, and so on.